+++ x-pack/plugin/ccr/src/main/java/org/elasticsearch/xpack/ccr/rest/RestUnfollowIndexAction.java
@@ -14,8 +14,8 @@ import org.elasticsearch.rest.action.RestToXContentListener;
 
 import java.io.IOException;
 
-import static org.elasticsearch.xpack.ccr.action.UnfollowIndexAction.INSTANCE;
-import static org.elasticsearch.xpack.ccr.action.UnfollowIndexAction.Request;
+import static org.elasticsearch.xpack.core.ccr.action.UnfollowIndexAction.INSTANCE;
+import static org.elasticsearch.xpack.core.ccr.action.UnfollowIndexAction.Request;
 
 public class RestUnfollowIndexAction extends BaseRestHandler {
